backend-ecommerce-v9

+++ src/controllers/cart.controllers.js
@@ -4,9 +4,9 @@ const Product = require('../models/Product');
 const Category = require('../models/Category');
 
 const getAll = catchError(async(req, res) => {
-    const { id } = req.user
+    const userId = req.user.id
     const results = await Cart.findAll({
-        where: {userId: id},
+        where: {userId},
         include:[
             {
                 model: Product,
@@ -39,27 +39,46 @@ const create = catchError(async(req, res) => {
 
 const getOne = catchError(async(req, res) => {
     const { id } = req.params;
-    const result = await Cart.findByPk(id);
+
+    const userId = req.user.id
+
+    const result = await Cart.findByPk(id, {
+        where: {userId},
+        include:[
+            {
+                model: Product,
+                attributes:{ exclude: ['createdAt', 'updatedAt'] },
+                include:[
+                    {
+                        model: Category,
+                        attributes:['name']
+                    }
+                ]
+            }
+        ]
+    });
     if(!result) return res.sendStatus(404);
     return res.json(result);
 });
 
 const remove = catchError(async(req, res) => {
     const { id } = req.params;
-    const result = await Cart.destroy({ where: {id} });
+    const userId = req.user.id
+    const result = await Cart.destroy({ where: {id, userId} });
     if(!result) return res.sendStatus(404);
     return res.sendStatus(204);
 });
 
 const update = catchError(async(req, res) => {
     const { id } = req.params;
+    const userId = req.user.id
 
     delete req.body.userId
     delete req.body.productId
 
     const result = await Cart.update(
         req.body,
-        { where: {id}, returning: true }
+        { where: {id, userId}, returning: true }
     );
     if(result[0] === 0) return res.sendStatus(404);
     return res.json(result[1][0]);
